ex 54

+++ week6/ex55/strings/strings1.cc
@@ -6,7 +6,6 @@ Strings::Strings(int argc, char **argv)
 	d_capacity = 0;
 	reserve((size_t) argc);
 
-	for (char **tmp = argv; tmp != argv + argc; ++tmp)
+	for (char **tmp = argv; tmp != argv + (size_t) argc; ++tmp)
 		add(*tmp);
-
 }
